requirements.txt
alembic==0.9.1
asn1crypto==0.24.0
bcrypt==3.1.4
bleach==2.0.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.2.2
dominate==2.3.1
et-xmlfile==1.0.1
Flask==0.12
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.0
Flask-Login==0.4.0
Flask-Migrate==2.0.3
Flask-Moment==0.5.1
Flask-PageDown==0.2.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
ForgeryPy==0.1
gunicorn==19.8.1
html5lib==0.999999999
idna==2.6
invoke==1.0.0
itsdangerous==0.24
jdcal==1.3
Jinja2==2.9.5
Mako==1.0.6
Markdown==2.6.8
MarkupSafe==0.23
olefile==0.44
openpyxl==2.5.2
paramiko==2.4.1
Pillow==4.0.0
pipenv==2018.6.25
pyasn1==0.4.3
pycparser==2.18
PyNaCl==1.2.1
pyperclip==1.6.0
pypinyin==0.16.1
python-editor==1.0.3
six==1.10.0
SQLAlchemy==1.1.5
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.3.0
visitor==0.1.3
webencodings==0.5
Werkzeug==0.11.15
WTForms==2.1
